Fixed function __calculateDestSrcXY()
This was difficult to debug because the error only occurred occasionally.
Example: Undefined array key "833"

The function for calculating the source and destination coordinates for
cropped images now uses separate index variables ($ix and $iy) instead
of $src_x / $src_y as array keys.

This prevents error messages that could occur depending on the original
image and crop dimensions.